// frontend/src/api/model/healthPlanCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthPlanCreate {
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  code: string;
}