// frontend/src/api/model/payerCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export interface PayerCreate {
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 30
     */
  code: string;
  tax_id?: string | null;
}